requirements.txt
numpy~=1.21.5
matplotlib~=3.5.2
scipy~=1.9.1
scikit-image~=0.19.2
ipython~=7.31.1
tensorboardx~=2.5.1
pillow~=9.2.0
hiddenlayer~=0.3
torchviz~=0.0.2
torchsummary~=1.5.1
requests~=2.28.1
pathlib~=1.0.1
tqdm~=4.64.1
setuptools~=63.4.1
tifffile~=2021.7.2
scikit-learn~=1.0.2
dataclasses~=0.8